+++ R/Spectra.R
@@ -365,15 +365,16 @@ NULL
 #'   the MS level specified with argument `msLevel`. Returns the filtered
 #'   `Spectra` (with spectra in their original order).
 #'
-#' - `filterMzRange`: filters the object keeping only peaks in each spectrum
-#'   that are within the provided m/z range.
+#' - `filterMzRange`: filters the object keeping or removing peaks in each
+#'   spectrum that are within the provided m/z range. Whether peaks are
+#'   retained or removed can be configured with parameter `keep` (default
+#'   `keep = TRUE`).
 #'
-#' - `filterMzValues`: filters the object keeping only peaks in each spectrum
-#'   that match the provided m/z value(s) (if parameter `keep = TRUE`, the
-#'   default) or removing them (if parameter `keep = FALSE`). The m/z matching
-#'   considers also the absolute `tolerance` and m/z-relative `ppm` values.
-#'   `tolerance` and `ppm` can be either of length 1 or equal to the length of
-#'   `mz` to define a different matching tolerance for each provided m/z.
+#' - `filterMzValues`: filters the object keeping **all** peaks in each
+#'   spectrum that match the provided m/z value(s) (for `keep = TRUE`, the
+#'   default) or removing **all** of them (for `keep = FALSE`). The m/z
+#'   matching considers also the absolute `tolerance` and m/z-relative
+#'   `ppm` values. `tolerance` and `ppm` have to be of length 1.
 #'
 #' - `filterPolarity`: filters the object keeping only spectra matching the
 #'   provided polarity. Returns the filtered `Spectra` (with spectra in their
@@ -773,9 +774,9 @@ NULL
 #' @param k For `pickPeaks`: `integer(1)`, number of values left and right of
 #'  the peak that should be considered in the weighted mean calculation.
 #'
-#' @param keep For `filterMzValues`: `logical(1)` whether the matching peaks
-#'     should be retained (`keep = TRUE`, the default`) or dropped
-#'     (`keep = FALSE`).
+#' @param keep For `filterMzValues` and `filterMzRange`: `logical(1)` whether
+#'     the matching peaks should be retained (`keep = TRUE`, the default`)
+#'     or dropped (`keep = FALSE`).
 #'
 #' @param keepIsotopes For `filterFourierTransformArtefacts`: whether isotope
 #'     peaks should not be removed as fourier artefacts.
@@ -862,9 +863,9 @@ NULL
 #'     should be passed along to the function defined with `FUN`. See function
 #'     description for details.
 #'
-#' @param tolerance For `compareSpectra`, `containsMz`: `numeric(1)` allowing to
-#'     define a constant maximal accepted difference between m/z values for
-#'     peaks to be matched. For `containsMz` and `filterMzValues` it can also
+#' @param tolerance For `compareSpectra`, `containsMz` and `filterMzValues`:
+#'     `numeric(1)` allowing to define a constant maximal accepted difference
+#'     between m/z values for peaks to be matched. For `containsMz` it can also
 #'     be of length equal `mz` to specify a different tolerance for each m/z
 #'     value.
 #'
@@ -1076,6 +1077,9 @@ NULL
 #' sps_sub <- filterMzValues(data, mz = c(103, 104),
 #'     tolerance = 0.3, keep = FALSE)
 #' mz(sps_sub)
+#'
+#' ## Note that `filterMzValues` keeps or removes all peaks with a matching
+#' ## m/z given the provided `ppm` and `tolerance` parameters.
 #'
 #' ## Filter a Spectra keeping only peaks within a m/z range
 #' sps_sub <- filterMzRange(data, mz = c(100, 300))
@@ -2059,17 +2063,20 @@ setMethod("filterMsLevel", "Spectra", function(object, msLevel. = integer()) {
 #'
 #' @export
 setMethod("filterMzRange", "Spectra",
-          function(object, mz = numeric(), msLevel. = uniqueMsLevels(object)) {
+          function(object, mz = numeric(), msLevel. = uniqueMsLevels(object),
+                   keep = TRUE) {
               if (!.check_ms_level(object, msLevel.))
                   return(object)
               if (!length(mz)) mz <- c(-Inf, Inf)
               else mz <- range(mz)
-              object <- addProcessing(object, .peaks_filter_mz_range,
-                                      mz = mz, msLevel = msLevel.,
+              object <- addProcessing(object, .peaks_filter_mz_range, mz = mz,
+                                      msLevel = msLevel., keep = keep,
                                       spectraVariables = "msLevel")
+              if (keep) keep_or_remove <- "select"
+              else keep_or_remove <- "remove"
               object@processing <- .logging(
-                  object@processing, "Filter: select peaks with an m/z within ",
-                  "[", mz[1L], ", ", mz[2L], "]")
+                  object@processing, "Filter: ", keep_or_remove,
+                  " peaks with an m/z within [", mz[1L], ", ", mz[2L], "]")
               object
           })
 
@@ -2082,10 +2089,10 @@ setMethod("filterMzValues", "Spectra",
               if (!.check_ms_level(object, msLevel.))
                   return(object)
               l <- length(mz)
-              if (!(length(tolerance) == 1 || length(tolerance) == l))
-                  stop("'tolerance' should be of length 1 or equal 'length(mz)'")
-              if (!(length(ppm) == 1 || length(ppm) == l))
-                  stop("'ppm' should be of length 1 or equal 'length(mz)'")
+              if (length(tolerance) != 1)
+                  stop("'tolerance' should be of length 1")
+              if (length(ppm) != 1)
+                  stop("'ppm' should be of length 1")
               if (is.unsorted(mz)) {
                   idx <- order(mz)
                   mz <- mz[idx]
@@ -2150,7 +2157,7 @@ setMethod("filterPrecursorMzRange", "Spectra",
 setMethod("filterPrecursorMzValues", "Spectra",
           function(object, mz = numeric(), ppm = 20, tolerance = 0) {
               object@backend <- filterPrecursorMzValues(
-                  object@backend, mz, ppm = ppm, tolerance = tolerance)
+                  object@backend, sort(mz), ppm = ppm, tolerance = tolerance)
               object@processing <- .logging(
                   object@processing,
                   "Filter: select spectra with precursor m/z matching ",