+++ NEWS.md
@@ -1,5 +1,13 @@
 # Spectra 1.9
 
+## Changes in 1.9.14
+
+- Fix issue with `filterMzValues` that would only keep (or remove) the first
+  matching peak instead of all matching peaks (given `ppm` and `tolerance`).
+  Issue [#274](https://github.com/rformassspectrometry/Spectra/issues/274).
+- Add parameter `keep` to `filterMzRange` to support keeping or removing
+  matching peaks.
+
 ## Changes in 1.9.13
 
 - Add the `backendBpparam` method that allows to evaluate whether a `MsBackend`